# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Event Volunteer Coordinator API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Volunteer {
  id: string;
  name: string;
  email: string;
  workstream?: string | null;
  is_admin: boolean;
  created_at: string;
}

export interface VolunteerUpdate {
  workstream?: string | null;
  is_admin?: boolean;
}

export interface RosterUploadResult {
  inserted: number;
  skipped: number;
  errors: string[];
}

export interface DeleteResult {
  success: boolean;
}

export interface SetupResult {
  success: boolean;
  message: string;
}

export interface ChannelSummary {
  id: string;
  name: string;
  slug: string;
  is_announcements: boolean;
  open_issues: number;
  open_urgents: number;
  latest_message_at?: string | null;
  latest_message_preview?: string | null;
}

export type AlertStatus = typeof AlertStatus[keyof typeof AlertStatus];


export const AlertStatus = {
  active: 'active',
  acknowledged: 'acknowledged',
  resolved: 'resolved',
} as const;

export interface Alert {
  id: string;
  channel_id: string;
  channel_name: string;
  user_id: string;
  sender_name: string;
  note: string;
  status: AlertStatus;
  acknowledged_by?: string | null;
  acknowledged_by_name?: string | null;
  acknowledged_at?: string | null;
  resolved_by?: string | null;
  resolved_at?: string | null;
  created_at: string;
}

export interface AgendaItem {
  id: string;
  sort_order: number;
  start_time: string;
  end_time: string;
  label: string;
  title: string;
  location: string;
  is_breakout: boolean;
  created_at: string;
}

export interface AgendaItemWrite {
  sort_order?: number;
  start_time?: string;
  end_time?: string;
  label?: string;
  title?: string;
  location?: string;
  is_breakout?: boolean;
}

export interface BreakoutSessionRow {
  id: string;
  track_id: string;
  sort_order: number;
  zone?: string | null;
  title: string;
  start_time: string;
  end_time: string;
  time_label: string;
  created_at: string;
}

export interface BreakoutSessionWrite {
  track_id?: string;
  sort_order?: number;
  zone?: string | null;
  title?: string;
  start_time?: string;
  end_time?: string;
  time_label?: string;
}

export interface BreakoutTrackRow {
  id: string;
  slug: string;
  sort_order: number;
  name: string;
  location: string;
  color: string;
  text_color: string;
  border_color: string;
  created_at: string;
}

export interface BreakoutTrackWrite {
  slug?: string;
  sort_order?: number;
  name?: string;
  location?: string;
  color?: string;
  text_color?: string;
  border_color?: string;
}

export interface BreakoutTrack {
  id: string;
  slug: string;
  sort_order: number;
  name: string;
  location: string;
  color: string;
  text_color: string;
  border_color: string;
  created_at: string;
  sessions: BreakoutSessionRow[];
}

export type ListVolunteersParams = {
workstream?: string;
is_admin?: boolean;
};