# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title — it controls generated filenames
  title: Api
  version: 0.1.0
  description: Event Volunteer Coordinator API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: admin
    description: Admin operations
  - name: channels
    description: Channel summaries
  - name: alerts
    description: Emergency alerts
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /admin/setup:
    post:
      operationId: setupDatabase
      summary: Initialize Supabase schema and seed channels
      tags: [admin]
      responses:
        "200":
          description: Setup complete
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SetupResult"

  /admin/roster:
    get:
      operationId: listVolunteers
      summary: List all volunteers
      tags: [admin]
      parameters:
        - name: workstream
          in: query
          schema:
            type: string
        - name: is_admin
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of volunteers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Volunteer"

  /admin/volunteers/{id}:
    patch:
      operationId: updateVolunteer
      summary: Update a volunteer workstream or admin status
      tags: [admin]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VolunteerUpdate"
      responses:
        "200":
          description: Updated volunteer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Volunteer"
    delete:
      operationId: deleteVolunteer
      summary: Remove a volunteer from the roster
      tags: [admin]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResult"

  /channels/summary:
    get:
      operationId: getChannelsSummary
      summary: All channels with open issue counts and latest message
      tags: [channels]
      responses:
        "200":
          description: Channel summaries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChannelSummary"

  /alerts/active:
    get:
      operationId: getActiveAlerts
      summary: Active and acknowledged alerts across all channels
      tags: [alerts]
      responses:
        "200":
          description: Active alerts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Alert"

  /agenda:
    get:
      operationId: getAgenda
      summary: List all agenda items ordered by sort_order
      tags: [schedule]
      responses:
        "200":
          description: Agenda items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AgendaItem"
    post:
      operationId: createAgendaItem
      summary: Create a new agenda item (admin only)
      tags: [schedule]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgendaItemWrite"
      responses:
        "201":
          description: Created agenda item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgendaItem"

  /agenda/{id}:
    patch:
      operationId: updateAgendaItem
      summary: Update an agenda item (admin only)
      tags: [schedule]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgendaItemWrite"
      responses:
        "200":
          description: Updated agenda item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgendaItem"
    delete:
      operationId: deleteAgendaItem
      summary: Delete an agenda item (admin only)
      tags: [schedule]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResult"

  /breakouts:
    get:
      operationId: getBreakouts
      summary: List all breakout tracks with their sessions
      tags: [schedule]
      responses:
        "200":
          description: Breakout tracks with sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BreakoutTrack"

  /breakouts/tracks:
    post:
      operationId: createBreakoutTrack
      summary: Create a breakout track (admin only)
      tags: [schedule]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BreakoutTrackWrite"
      responses:
        "201":
          description: Created track
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BreakoutTrackRow"

  /breakouts/tracks/{id}:
    patch:
      operationId: updateBreakoutTrack
      summary: Update a breakout track (admin only)
      tags: [schedule]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BreakoutTrackWrite"
      responses:
        "200":
          description: Updated track
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BreakoutTrackRow"
    delete:
      operationId: deleteBreakoutTrack
      summary: Delete a breakout track and its sessions (admin only)
      tags: [schedule]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResult"

  /breakouts/sessions:
    post:
      operationId: createBreakoutSession
      summary: Create a breakout session (admin only)
      tags: [schedule]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BreakoutSessionWrite"
      responses:
        "201":
          description: Created session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BreakoutSessionRow"

  /breakouts/sessions/{id}:
    patch:
      operationId: updateBreakoutSession
      summary: Update a breakout session (admin only)
      tags: [schedule]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BreakoutSessionWrite"
      responses:
        "200":
          description: Updated session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BreakoutSessionRow"
    delete:
      operationId: deleteBreakoutSession
      summary: Delete a breakout session (admin only)
      tags: [schedule]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResult"

components:
  schemas:
    HealthStatus:
      type: object
      required: [status]
      properties:
        status:
          type: string

    Volunteer:
      type: object
      required: [id, name, email, is_admin, created_at]
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        workstream:
          type: string
          nullable: true
        is_admin:
          type: boolean
        created_at:
          type: string
          format: date-time

    VolunteerUpdate:
      type: object
      properties:
        workstream:
          type: string
          nullable: true
        is_admin:
          type: boolean

    RosterUploadResult:
      type: object
      required: [inserted, skipped, errors]
      properties:
        inserted:
          type: integer
        skipped:
          type: integer
        errors:
          type: array
          items:
            type: string

    DeleteResult:
      type: object
      required: [success]
      properties:
        success:
          type: boolean

    SetupResult:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
        message:
          type: string

    ChannelSummary:
      type: object
      required: [id, name, slug, is_announcements, open_issues, open_urgents]
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string
        is_announcements:
          type: boolean
        open_issues:
          type: integer
        open_urgents:
          type: integer
        latest_message_at:
          type: string
          format: date-time
          nullable: true
        latest_message_preview:
          type: string
          nullable: true

    Alert:
      type: object
      required: [id, channel_id, channel_name, user_id, sender_name, note, status, created_at]
      properties:
        id:
          type: string
        channel_id:
          type: string
        channel_name:
          type: string
        user_id:
          type: string
        sender_name:
          type: string
        note:
          type: string
        status:
          type: string
          enum: [active, acknowledged, resolved]
        acknowledged_by:
          type: string
          nullable: true
        acknowledged_by_name:
          type: string
          nullable: true
        acknowledged_at:
          type: string
          format: date-time
          nullable: true
        resolved_by:
          type: string
          nullable: true
        resolved_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time

    AgendaItem:
      type: object
      required: [id, sort_order, start_time, end_time, label, title, location, is_breakout, created_at]
      properties:
        id:
          type: string
        sort_order:
          type: integer
        start_time:
          type: string
        end_time:
          type: string
        label:
          type: string
        title:
          type: string
        location:
          type: string
        is_breakout:
          type: boolean
        created_at:
          type: string
          format: date-time

    AgendaItemWrite:
      type: object
      properties:
        sort_order:
          type: integer
        start_time:
          type: string
        end_time:
          type: string
        label:
          type: string
        title:
          type: string
        location:
          type: string
        is_breakout:
          type: boolean

    BreakoutSessionRow:
      type: object
      required: [id, track_id, sort_order, title, start_time, end_time, time_label, created_at]
      properties:
        id:
          type: string
        track_id:
          type: string
        sort_order:
          type: integer
        zone:
          type: string
          nullable: true
        title:
          type: string
        start_time:
          type: string
        end_time:
          type: string
        time_label:
          type: string
        created_at:
          type: string
          format: date-time

    BreakoutSessionWrite:
      type: object
      properties:
        track_id:
          type: string
        sort_order:
          type: integer
        zone:
          type: string
          nullable: true
        title:
          type: string
        start_time:
          type: string
        end_time:
          type: string
        time_label:
          type: string

    BreakoutTrackRow:
      type: object
      required: [id, slug, sort_order, name, location, color, text_color, border_color, created_at]
      properties:
        id:
          type: string
        slug:
          type: string
        sort_order:
          type: integer
        name:
          type: string
        location:
          type: string
        color:
          type: string
        text_color:
          type: string
        border_color:
          type: string
        created_at:
          type: string
          format: date-time

    BreakoutTrackWrite:
      type: object
      properties:
        slug:
          type: string
        sort_order:
          type: integer
        name:
          type: string
        location:
          type: string
        color:
          type: string
        text_color:
          type: string
        border_color:
          type: string

    BreakoutTrack:
      type: object
      required: [id, slug, sort_order, name, location, color, text_color, border_color, created_at, sessions]
      properties:
        id:
          type: string
        slug:
          type: string
        sort_order:
          type: integer
        name:
          type: string
        location:
          type: string
        color:
          type: string
        text_color:
          type: string
        border_color:
          type: string
        created_at:
          type: string
          format: date-time
        sessions:
          type: array
          items:
            $ref: "#/components/schemas/BreakoutSessionRow"