# lib/api-zod/src/generated/types/agendaItemWrite.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Event Volunteer Coordinator API
 * OpenAPI spec version: 0.1.0
 */

export interface AgendaItemWrite {
  sort_order?: number;
  start_time?: string;
  end_time?: string;
  label?: string;
  title?: string;
  location?: string;
  is_breakout?: boolean;
}