# lib/api-zod/src/generated/types/breakoutSessionRow.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Event Volunteer Coordinator API
 * OpenAPI spec version: 0.1.0
 */

export interface BreakoutSessionRow {
  id: string;
  track_id: string;
  sort_order: number;
  zone?: string | null;
  title: string;
  start_time: string;
  end_time: string;
  time_label: string;
  created_at: Date;
}