# lib/api-zod/src/generated/types/breakoutTrack.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Event Volunteer Coordinator API
 * OpenAPI spec version: 0.1.0
 */
import type { BreakoutSessionRow } from './breakoutSessionRow';

export interface BreakoutTrack {
  id: string;
  slug: string;
  sort_order: number;
  name: string;
  location: string;
  color: string;
  text_color: string;
  border_color: string;
  created_at: Date;
  sessions: BreakoutSessionRow[];
}