# lib/api-zod/src/generated/types/breakoutTrackRow.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Event Volunteer Coordinator API
 * OpenAPI spec version: 0.1.0
 */

export interface BreakoutTrackRow {
  id: string;
  slug: string;
  sort_order: number;
  name: string;
  location: string;
  color: string;
  text_color: string;
  border_color: string;
  created_at: Date;
}