# lib/api-zod/src/generated/types/breakoutTrackWrite.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Event Volunteer Coordinator API
 * OpenAPI spec version: 0.1.0
 */

export interface BreakoutTrackWrite {
  slug?: string;
  sort_order?: number;
  name?: string;
  location?: string;
  color?: string;
  text_color?: string;
  border_color?: string;
}